fix: refactor date context
